another commit

+++ src/q8.java
@@ -6,13 +6,13 @@ public class q8 {
     public static void main(String[] args){
         Scanner input = new Scanner(System.in);
         System.out.print("Enter the number (n) of numbers you want to calculate the average: ");
-        int number = input.nextInt();
-        int sum = 0;
+        float number = input.nextFloat();
+        float sum = 0;
         int x = 0;
 
         while (number > x){
             System.out.print("Enter number: ");
-            sum += input.nextInt();
+            sum += input.nextFloat();
             x += 1;
         }
